Added Metric Conversion To Vehicle,Driver Performance,Vehicle Performance, Driver Style, and Vehicle Style PDF report.

+++ dao/src/main/java/com/inthinc/pro/dao/util/MeasurementConversionUtil.java
@@ -40,7 +40,7 @@ public class MeasurementConversionUtil
         return bigDecimal.movePointLeft(1).floatValue();
     }
 
-    public static Float convertDistance(Integer distance, MeasurementType convertToMeasurementType)
+    public static Float convertDistance(Float distance, MeasurementType convertToMeasurementType)
     {
         if (convertToMeasurementType == MeasurementType.METRIC)
         {
